Consolidate file system utils into the same file
